Refactor building list menu according to concordia_constants changes Refactor building model into concordia building model Fix camera position hiding building below info sheet
