+++ lib/blocs/search_bloc/search_state.dart
@@ -1,4 +1,4 @@
-import 'package:concordia_go/utilities/concordia_constants.dart';
+import 'package:concordia_go/models/concordia_building_model.dart';
 import 'package:meta/meta.dart';
 
 @immutable
