--- lib/models/building_model.dart
@@ -1,10 +1,0 @@
-import 'package:google_maps_flutter/google_maps_flutter.dart';
-
-class Building {
-  String code;
-  String name;
-  LatLng coordinates;
-  String address;
-  String campus;
-  Map<String, String> hours;
-}
