+++ lib/models/concordia_building_model.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/foundation.dart';
+import 'package:google_maps_flutter/google_maps_flutter.dart';
+
+enum Campus { SGW, Loyola }
+
+class ConcordiaBuilding {
+  final String code;
+  final String name;
+  final String address;
+  final Campus campus;
+  final LatLng coordinates;
+  final List<LatLng> vertices;
+  final Map<String, String> hours;
+
+  const ConcordiaBuilding({
+    this.code,
+    this.name,
+    this.address,
+    this.campus,
+    this.coordinates,
+    this.vertices,
+    this.hours,
+  });
+
+  String campusString() {
+    String campusName = describeEnum(campus);
+    return '${campusName} Campus';
+  }
+}
+
+String campusString(Campus campus) {
+  String campusName = describeEnum(campus);
+  return '${campusName} Campus';
+}
